colors separated as variables in styles/colors.scss
