+++ package.json
@@ -17,6 +17,7 @@
     "react-infinite-scroll-component": "^6.1.0",
     "react-router-dom": "^6.8.1",
     "react-scripts": "5.0.1",
+    "recharts": "^2.4.3",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0"
   },
