+++ src/pages/Market/components/Input/Input.module.scss
@@ -1,27 +1,29 @@
+@import "@styles/colors.scss";
+
 .input {
-  background: #ffffff;
-  border: 1px solid #dfe2e4;
-  border-radius: 4px;
-  padding: 12px;
-  font-family: 'Circular Std';
-  font-style: normal;
-  font-weight: 450;
-  font-size: 12px;
-  line-height: 15px;
-  color: #343a40;
-  box-sizing: border-box;
-  width: 100%;
+	background: #ffffff;
+	border: 1px solid $border;
+	border-radius: 4px;
+	padding: 12px;
+	font-family: "Circular Std";
+	font-style: normal;
+	font-weight: 450;
+	font-size: 12px;
+	line-height: 15px;
+	color: #343a40;
+	box-sizing: border-box;
+	width: 100%;
 
-  &::placeholder {
-    color: #dfe2e4;
-  }
+	&::placeholder {
+		color: $border;
+	}
 
-  &:focus {
-    border-color: #6c757d;
-  }
+	&:focus {
+		border-color: $text_secondary;
+	}
 
-  &:disabled {
-    color: #6c757d;
-    // background: #f8f9fa;
-  }
+	&:disabled {
+		color: $text_secondary;
+		// background: #f8f9fa;
+	}
 }
