+++ src/styles/colors.scss
@@ -1,0 +1,6 @@
+$background: #f8f9fa;
+$text: #212529;
+$text_secondary: #6c757d;
+$green: #21bf73;
+$red: red;
+$border: #dfe2e4;
