+++ src/styles/styles.scss
@@ -1,6 +1,5 @@
 @import "@assets/fonts/CircularStd/fonts";
 @import "./reset";
-@import "./colors";
 
 html {
 	font-family: "Circular Std";
